app/requirements.txt
snowplow-tracker ~> 0.8.0
psycopg2-binary
jsonschema